+++ sdk/securitydevops/azure-resourcemanager-securitydevops/CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## 1.0.0-beta.1 (2022-10-13)
 
+- Azure Resource Manager Security Dev Ops client library for Java. This package contains Microsoft Azure SDK for Security Dev Ops Management SDK. REST APIs for Defender for DevOps. Package tag package-2022-09-01-preview. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 - Azure Resource Manager securitydevops client library for Java. This package contains Microsoft Azure SDK for securitydevops Management SDK. REST APIs for Defender for DevOps. Package tag package-2022-09-01-preview. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
+